+++ src/components/Login.vue
@@ -74,7 +74,7 @@ export default {
        DataPostApi.validateLogin(username,password)
       .then(response => {
         if(response.isLoginSuccess === true) {
-          this.$router.push("/dashboard");
+          this.$router.push({ name: "dashboard", params: {username: username } });
         } else {
            this.errors.push('Invalid Credentials');
         }
